Maze Generator And Solver
The application has two modes. In the first, it makes the map by pulling information from a website containing matrix information. In the other, it randomly creates the map. Then, in the first mode, you can turn off the screen and solve it without seeing the road, or it has the ability to solve it by seeing the road if you want.